Extract Callbacks to a separate source file
