+++ src/debot/pipechain.rs
@@ -14,7 +14,7 @@ pub enum ApproveKind {
 
 #[derive(Deserialize, Clone, Default)]
 #[serde(rename_all(deserialize = "camelCase"))]
-pub struct DebotManifest {
+pub struct PipeChain {
     pub version: u8,
     pub debot_address: String,
     #[serde(default = "default_init_method")]
